Unify Observable/Single implementations
