+++ rxlifecycle/src/main/java/com/trello/rxlifecycle/UntilCorrespondingEventObservableTransformer.java
@@ -1,9 +1,9 @@
 package com.trello.rxlifecycle;
 
 import rx.Observable;
-import rx.exceptions.Exceptions;
 import rx.functions.Func1;
-import rx.functions.Func2;
+
+import static com.trello.rxlifecycle.TakeUntilGenerator.takeUntilCorrespondingEvent;
 
 /**
  * Continues a subscription until it sees a particular lifecycle event.
@@ -23,37 +23,6 @@ class UntilCorrespondingEventObservableTransformer<T, R> implements Observable.T
 
     @Override
     public Observable<T> call(Observable<T> source) {
-        return source.takeUntil(
-            Observable.combineLatest(
-                this.sharedLifecycle.take(1).map(correspondingEvents),
-                this.sharedLifecycle.skip(1),
-                new Func2<R, R, Boolean>() {
-                    @Override
-                    public Boolean call(R bindUntilEvent, R lifecycleEvent) {
-                        return lifecycleEvent.equals(bindUntilEvent);
-                    }
-                })
-                .onErrorReturn(RESUME_FUNCTION)
-                .takeFirst(SHOULD_COMPLETE)
-        );
+        return source.takeUntil(takeUntilCorrespondingEvent(sharedLifecycle, correspondingEvents));
     }
-
-    private static final Func1<Throwable, Boolean> RESUME_FUNCTION = new Func1<Throwable, Boolean>() {
-        @Override
-        public Boolean call(Throwable throwable) {
-            if (throwable instanceof OutsideLifecycleException) {
-                return true;
-            }
-
-            Exceptions.propagate(throwable);
-            return false;
-        }
-    };
-
-    private static final Func1<Boolean, Boolean> SHOULD_COMPLETE = new Func1<Boolean, Boolean>() {
-        @Override
-        public Boolean call(Boolean shouldComplete) {
-            return shouldComplete;
-        }
-    };
 }
