+++ rxlifecycle/src/main/java/com/trello/rxlifecycle/UntilEventSingleTransformer.java
@@ -2,7 +2,8 @@ package com.trello.rxlifecycle;
 
 import rx.Observable;
 import rx.Single;
-import rx.functions.Func1;
+
+import static com.trello.rxlifecycle.TakeUntilGenerator.takeUntilEvent;
 
 /**
  * Continues a subscription until it sees a particular lifecycle event.
@@ -19,13 +20,6 @@ class UntilEventSingleTransformer<T, R> implements Single.Transformer<T, T> {
 
     @Override
     public Single<T> call(Single<T> source) {
-        return source.takeUntil(
-            lifecycle.takeFirst(new Func1<R, Boolean>() {
-                @Override
-                public Boolean call(R lifecycleEvent) {
-                    return lifecycleEvent.equals(event);
-                }
-            })
-        );
+        return source.takeUntil(takeUntilEvent(lifecycle, event));
     }
 }
